Added Switch to Change Kcmax Calc Based on Reference Crop
I added io.rfcrp as a variable and a conditional statement to change the method of calculing Kcmax based on whether the short or tall reference crop is used. My initial edit contains an else statement that gives an error if the user uses a reference crop other than "S" or "T". However, I do not think the error statement would actually arise in any real use case, so I will take it out.

My initial commit also contained a typo in the header.