8_23_2022 - Reworked Logic in the TAW Switch
In my previous commits, the calculation of TAW for the dynamic root zone was not correct - it overestimated TAW. I went back and reworked the logic of the TAW calculation to operate in a similar (but more flexible) way to the LIRF WaterBalance spreadsheet. All of the major edits came in the second for-loop of the TAW calculation block. The current operation of the loop is dependent on the most recent version of the SoilWater class (see commit "8_23_2022 - Altered SWC to Support Model Changes" on develop-soil-water-class branch).