8_25_2022 - SoilWater Class + pyfao56 Demo Files
I am committing the files that I use to run pyfao56 with the new SoilWater Class as a way to show how I intend it to work. The main file I use to run things at the moment is pyfao56-LIRF_Integration_Testing.py. I am also committing some output files to show the way in which it works. The customload Weather child class I currently use is a mess, but it works.