8_29_2022 - First Draft of SWC Changes per Thorp's Suggestions
I have what I think is a working draft of the changes to the SoilWater class that Kelly asked for. I have replaced the "create_soil_water_profile_" method with the class instantiation. I also created an input file design for the class to read and write to. I added the savefile and loadfile methods in support of that new design element. I also converted the soil_water_profile object from a dictionary to a dataframe and refactored the usages in the model to reflect that conversion.

Moving forward:
- I am going to move the root zone features to a class of their own.
- I need to reorganize / clean-up the supporting documentation of the class. I hope to have a commit with each of those in the future.