8_30_2022 - Final Draft of SWC Changes per Thorp's Suggestions
Final draft of the revisions to the SoilWater class that were requested by Kelly Thorp via email. In these revisions, I substantially reworked the SoilWater class by:
- moving the create_soil_water_profile method's functionality to the class initialization method,
- creating an input file design for the class to read and write to,
- adding savefile and loadfile methods,
- making the __str__ method format more similar to other pyfao56 class __str__ methods,
- replacing the soil_water_profile dictionary with a Pandas dataframe of the same name,
- and updating model class as well as all supporting documentation to reflect these changes.

The new file titled "testing_integration8_29_22.py" was used primarily to test the savefile and loadfile methods. That file should be ready to run for anyone who would like to test out the new class themselves, but that file does not run the model. It simply uses the SoilWater class.