09_15_2022 - Line 393 HotFix
I reviewed the logic in the model once again, and I saw a potential issue/bug that the conditional statement that was on line 393 could cause. I removed that conditional because it was a conditional meant to avoid some of the logic in the (probably rare) case that the root zone happened to be deeper than the deepest soil layer provided. Such a conditional is extraneous to the core functionality (it only would increase efficiency a tiny bit by skipping approximately 10 lines of code), so it could be cut with ease.

I tested this change and it still works as expected. I wanted to get this simple alteration in before the branch was merged to develop.