09_21_2022 - SoilProfile Revisions & Example I/O Script
I revised the SoilProfile class based on Kelly's comments in his previous commit. I reworked the sdata DataFrame to only contain three columns and bottom-layer depths (cm, int) as the index.
I changed the __str__, savefile, and loadfile methods in support of the changes to the sdata DataFrame.
I also reread all of the docstrings and default file formats to make sure they are all up-to-date. Everything looks good to me now.
I did leave a couple notes at the top of the class for Kelly to review.

I also made a script to demonstrate the input/output functionality of the SoilProfile class. That script is in the test5 directory. There are also some example output files in that directory.