Made Zr a function of tabular Kcb
Updated the Zr equation in model.py such that Zr is always based on tabular Kcb, even when users give Kcb updated from the update file.

Also reran all of the tests, and the change altered the files you would expect it to alter. But, I am not sure what is up with the new spaces in AZMET_Maricopa_hourlyEToETr.txt. I did not change anything that should change that output, so I don't know what is going on there.