Simplified Root Zone Soil Water Loop
Kelly asked me to simplify the main loop in the compute_root_zone_sw method. Previously, it was a fairly complex loop that turned the swddata and swcdata dataframes into nested dictionaries, and then computed values from there. Now, it just uses some of the built-in Pandas methods to accomplish the same goal. Although the output is the same, the compute_root_zone_sw method seems to be much slower (at least on my machine) than it was when it was the more complex dictionary approach. So, there might be a speed/simplicity tradeoff to consider here.