Removed swddata Attribute & vswd File Methods and Fixed Loadfile Bug
I simplified the class by removing the swddata class attribute. As a result, I was also able to remove the .vswd I/O file methods, which greatly simplified the class.

While testing, I also discovered that the rzswd file was not being loaded correctly. I fixed that bug so that it now works as intended.