Fixed Empty Irrigation Class Issue
Previously, an empty irrigation file would break the model. I altered the irrigation.py str function such that empty classes create files with -999 values to keep the model from breaking. This is all part of the effort to make the irrigation class optional and guard against users giving unexpected inputs to the model.