Fixed Update Class NaN String
During in-season use, I noticed that the Update class NaN rep string was off a little bit. Figured out that issue.

Also added empty dataframe safeguard for users who make a file out of an empty update class. I just reused the code that was used for the same purpose in the irrigation class.