Adjusted Default hini value
I searched FAO-56 for more insight on what initial height should usually be assumed to be, and I couldn't find anything.

Kelly mentioned that the 0.05m value is probably a remanent of the Maricopa spreadsheets.

I changed the default value to 0.013m, which is roughly half an inch. I think that assumption is a little more realistic for most initial height, but I also don't think this parameter matters much at all.