Added one crucial line of code, comments, and ran tests
I added a line of code that removes a comma and makes this code work as intended.

I also added some block comments trying to explain the basic functionality of the code changes. I also ran all of the tests. All of the tests ran, although the output files changed. From what I can tell, the only change to the output files is the new Kcadj column.