Kcb adjustment as on page 136
Kcb adjustments were added based on the climate of the region using equation 70 on page  136 of FAO56.  In initial tests, Kcb adjusted values decreased for humid and increased for semi-arid as expected.
Changes are made only in the Model class.  The position of additional code chunk may need to adjust